Refactor Program into models and services
