+++ src/unifocl/Models/CliSessionState.cs
@@ -1,0 +1,6 @@
+internal sealed class CliSessionState
+{
+    public int? AttachedPort { get; set; }
+    public string? CurrentProjectPath { get; set; }
+    public DateTimeOffset? LastOpenedUtc { get; set; }
+}
