+++ src/unifocl/Services/ProcessUtil.cs
@@ -1,0 +1,17 @@
+using System.Diagnostics;
+
+internal static class ProcessUtil
+{
+    public static bool IsAlive(int pid)
+    {
+        try
+        {
+            var process = Process.GetProcessById(pid);
+            return !process.HasExited;
+        }
+        catch
+        {
+            return false;
+        }
+    }
+}
